backend-python/requirements.txt
# Core LangChain & LangGraph
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0

# LLM Providers
langchain-openai>=0.2.0
langchain-ollama>=0.2.0

# Embeddings & Vector Stores
langchain-huggingface>=0.1.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# Document Loaders
pypdf>=4.0.0
python-docx>=1.0.0
docx2txt>=0.8
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
aiofiles>=23.0.0
httpx>=0.26.0
